% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_CV.R
\name{predict.conformalCV}
\alias{predict.conformalCV}
\title{Predict Method for conformalCV objects}
\usage{
\method{predict}{conformalCV}(
  object,
  Xtest,
  alpha = 0.1,
  wthigh = 20,
  wtlow = 0.05,
  useInf = FALSE,
  ...
)
}
\arguments{
\item{object}{an object of class \code{conformalCV}; see \code{\link{conformal}}.}

\item{Xtest}{testing covariates.}

\item{alpha}{confidence level.}

\item{wthigh}{upper truncation level of weights; see Details.}

\item{wtlow}{lower truncation level of weights; see Details.}

\item{useInf}{if FALSE then replace infinity by the maximum conformity score.}

\item{...}{other arguments}
}
\value{
predictive intervals. A data.frame with \code{nrow(Xtest)} rows and two columns:
"lower" for the lower bound and "upper" for the upper bound.
}
\description{
Obtains predictive intervals on a testing dataset based on a \code{conformalCV} object
from \code{\link{conformal}} with \code{useCV = TRUE}.
}
\details{
Given a testing set \eqn{X_1, X_2, \ldots, X_n} and a weight function \eqn{w(x)}, the
weight of the weighted distribution \eqn{p_j = w(X_j) / (w(X_1) + \cdots + w(X_n))}.
In cases where some of \eqn{p_j} are extreme, we truncate \eqn{p_j} at level \code{wthigh}
and \code{wtlow} to ensure stability. If \code{wthigh = Inf, wtlow = 0}, no truncation
is being used.
}
\seealso{
\code{\link{predict.conformalSplit}}, \code{\link{conformal}}.
}
