% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformalInt_CV.R
\name{predict.conformalIntCV}
\alias{predict.conformalIntCV}
\title{Predict Method for conformalIntCV objects}
\usage{
\method{predict}{conformalIntCV}(object, Xtest, alpha = 0.1, wthigh = 20, wtlow = 0.05, ...)
}
\arguments{
\item{object}{an object of class \code{conformalIntCV}; see \code{\link{conformalInt}}.}

\item{Xtest}{testing covariates.}

\item{alpha}{confidence level.}

\item{wthigh}{upper truncation level of weights; see Details.}

\item{wtlow}{lower truncation level of weights; see Details.}

\item{...}{other arguments}
}
\value{
predictive intervals. A data.frame with \code{nrow(Xtest)} rows and two columns:
"lower" for the lower bound and "upper" for the upper bound.
}
\description{
Obtains predictive intervals on a testing dataset based on a \code{conformalIntCV} object
from \code{\link{conformalInt}} with \code{useCV = TRUE}.
}
\details{
Given a testing set \eqn{X_1, X_2, \ldots, X_n} and a weight function \eqn{w(x)}, the
weight of the weighted distribution \eqn{p_j = w(X_j) / (w(X_1) + \cdots + w(X_n))}.
In cases where some of \eqn{p_j} are extreme, we truncate \eqn{p_j} at level \code{wthigh}
and \code{wtlow} to ensure stability. If \code{wthigh = Inf, wtlow = 0}, no truncation
is being used.
}
\seealso{
\code{\link{predict.conformalIntSplit}}, \code{\link{conformalInt}}.
}
